pyproject.toml
[tool.poetry]
name = "distributed-automation-example"
version = "0.1.0"
description = ""
authors = ["Omid Mortazavi <omid_mortazavi@hotmail.com>"]
readme = "README.md"
packages = [{include = "distributed_automation_example"}]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
pyopenssl = "^23.2.0"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
six = "^1.16.0"
rich = "^13.4.2"
flake8 = "^6.0.0"
black = "^23.3.0"
nornir = "^3.3.0"
netmiko = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
